Fix include file resolution on Windows
Fix for failure to resolve include files on Windows when the files in
question are virtual, i.e. created with mapVirtualFile().
The files are searched for in a hash which is keyed on the file's
normalized path as a string. Searches for a file in that hash will fail
unless the search path is in the exact same textual form.